pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ReAXET", email = "joshua.sturm@outlook.com"},
]
dependencies = [
    "litestar>=2.6.1",
    "advanced-alchemy>=0.7.4",
    "fastapi[all]>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "sqlalchemy>=2.0.27",
    "starlette>=0.36.3",
    "gunicorn>=21.2.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "pre-commit>=3.6.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
